# backend/src/routes/auth.ts
/**
 * Authentication Routes
 * Login, register, logout, token refresh, password management
 */

import { Router, Request, Response } from 'express';
import { z } from 'zod';
import { authenticate, verifyRefreshToken } from '../middleware/auth';
import { logAuditEvent } from '../middleware/audit';

const router = Router();

// Validation schemas
const loginSchema = z.object({
  email: z.string().email('Invalid email address'),
  password: z.string().min(8, 'Password must be at least 8 characters'),
  mfaCode: z.string().length(6).optional(),
});

const registerSchema = z.object({
  email: z.string().email('Invalid email address'),
  password: z
    .string()
    .min(8, 'Password must be at least 8 characters')
    .regex(
      /^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]/,
      'Password must contain uppercase, lowercase, number, and special character'
    ),
  firstName: z.string().min(1, 'First name is required'),
  lastName: z.string().min(1, 'Last name is required'),
  role: z.enum(['patient', 'doctor', 'nurse']).default('patient'),
  organizationId: z.string().uuid().optional(),
});

const forgotPasswordSchema = z.object({
  email: z.string().email('Invalid email address'),
});

const resetPasswordSchema = z.object({
  token: z.string().min(1, 'Reset token is required'),
  password: z
    .string()
    .min(8)
    .regex(/^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]/),
});

const changePasswordSchema = z.object({
  currentPassword: z.string().min(1, 'Current password is required'),
  newPassword: z
    .string()
    .min(8)
    .regex(/^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]/),
});

/**
 * POST /auth/login
 * Authenticate user and return tokens
 */
router.post('/login', async (req: Request, res: Response) => {
  try {
    const data = loginSchema.parse(req.body);

    // TODO: Implement actual authentication
    // 1. Find user by email
    // 2. Verify password with bcrypt
    // 3. Check MFA if enabled
    // 4. Generate JWT and refresh token
    // 5. Set refresh token in httpOnly cookie
    // 6. Log audit event

    await logAuditEvent('LOGIN', {
      userEmail: data.email,
      ipAddress: req.ip,
      status: 'success',
    });

    res.json({
      status: 'success',
      data: {
        accessToken: 'jwt_token_here',
        expiresIn: 900, // 15 minutes
        user: {
          id: 'user-id',
          email: data.email,
          role: 'doctor',
          firstName: 'Demo',
          lastName: 'User',
        },
      },
    });
  } catch (error) {
    await logAuditEvent('LOGIN_FAILED', {
      userEmail: req.body?.email,
      ipAddress: req.ip,
      status: 'failure',
      errorMessage: error instanceof Error ? error.message : 'Unknown error',
    });

    throw error;
  }
});

/**
 * POST /auth/register
 * Register new user
 */
router.post('/register', async (req: Request, res: Response) => {
  const data = registerSchema.parse(req.body);

  // TODO: Implement registration
  // 1. Check if email already exists
  // 2. Hash password with bcrypt
  // 3. Create user record
  // 4. Send verification email
  // 5. Log audit event

  res.status(201).json({
    status: 'success',
    message: 'Registration successful. Please verify your email.',
    data: {
      userId: 'new-user-id',
    },
  });
});

/**
 * POST /auth/refresh
 * Refresh access token using refresh token
 */
router.post('/refresh', verifyRefreshToken, async (req: Request, res: Response) => {
  // TODO: Generate new access token from refresh token
  res.json({
    status: 'success',
    data: {
      accessToken: 'new_jwt_token',
      expiresIn: 900,
    },
  });
});

/**
 * POST /auth/logout
 * Logout user and invalidate tokens
 */
router.post('/logout', authenticate, async (req: Request, res: Response) => {
  // TODO: Invalidate refresh token, clear session

  await logAuditEvent('LOGOUT', {
    userId: req.user?.userId,
    ipAddress: req.ip,
  });

  res.clearCookie('refreshToken');
  res.json({
    status: 'success',
    message: 'Logged out successfully',
  });
});

/**
 * POST /auth/forgot-password
 * Request password reset
 */
router.post('/forgot-password', async (req: Request, res: Response) => {
  const data = forgotPasswordSchema.parse(req.body);

  // TODO: Generate reset token, send email
  // Always return success to prevent email enumeration

  res.json({
    status: 'success',
    message: 'If an account exists with this email, you will receive a password reset link.',
  });
});

/**
 * POST /auth/reset-password
 * Reset password with token
 */
router.post('/reset-password', async (req: Request, res: Response) => {
  const data = resetPasswordSchema.parse(req.body);

  // TODO: Verify token, update password

  await logAuditEvent('PASSWORD_RESET', {
    ipAddress: req.ip,
    status: 'success',
  });

  res.json({
    status: 'success',
    message: 'Password has been reset successfully',
  });
});

/**
 * POST /auth/change-password
 * Change password for authenticated user
 */
router.post('/change-password', authenticate, async (req: Request, res: Response) => {
  const data = changePasswordSchema.parse(req.body);

  // TODO: Verify current password, update to new password

  await logAuditEvent('PASSWORD_CHANGE', {
    userId: req.user?.userId,
    ipAddress: req.ip,
    status: 'success',
  });

  res.json({
    status: 'success',
    message: 'Password changed successfully',
  });
});

/**
 * GET /auth/me
 * Get current user profile
 */
router.get('/me', authenticate, async (req: Request, res: Response) => {
  // TODO: Fetch full user profile from database

  res.json({
    status: 'success',
    data: {
      id: req.user?.userId,
      email: req.user?.email,
      role: req.user?.role,
    },
  });
});

export default router;